requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.23.0
starlette==0.27.0
python-multipart==0.0.6
hvac==1.1.0